+++ aperag/views/auth.py
@@ -522,41 +522,3 @@ async def delete_user_view(
     await async_db_ops.delete_user(session, target)
     return {"message": "User deleted successfully"}
 
-
-@router.post("/test/register_admin", tags=["test"])
-async def test_register_admin(
-    request: Request, data: view_models.Register, session: AsyncSessionDep, x_test_secret: str = Header(None)
-):
-    if os.environ.get("TEST_ENV") != "1":
-        raise HTTPException(status_code=403, detail="Not allowed")
-
-    secret = os.environ.get("TEST_ADMIN_SECRET", "").strip()
-    if not secret or secret != x_test_secret:
-        raise HTTPException(status_code=403, detail="Not allowed")
-
-    from sqlalchemy import select
-
-    result = await session.execute(select(User).where(User.username == data.username))
-    if result.scalars().first():
-        raise HTTPException(status_code=400, detail="Username already exists")
-
-    user = User(
-        username=data.username,
-        email=data.email,
-        hashed_password=UserManager(None).password_helper.hash(data.password),
-        role=Role.ADMIN,
-        is_active=True,
-        is_verified=True,
-        date_joined=utc_now(),
-    )
-    session.add(user)
-    await session.commit()
-    await session.refresh(user)
-    return view_models.User(
-        id=str(user.id),
-        username=user.username,
-        email=user.email,
-        role=user.role,
-        is_active=user.is_active,
-        date_joined=user.date_joined.isoformat(),
-    )
