+++ tests/e2e_test/config.py
@@ -47,7 +47,4 @@ COMPLETION_MODEL_CUSTOM_PROVIDER = os.getenv("COMPLETION_MODEL_CUSTOM_PROVIDER",
 RERANK_MODEL_NAME = os.getenv("RERANK_MODEL_NAME", "gte-rerank-v2")
 RERANK_MODEL_CUSTOM_PROVIDER = os.getenv("RERANK_MODEL_CUSTOM_PROVIDER", "alibabacloud")
 
-TEST_ENV = os.getenv("TEST_ENV", "1")
-TEST_ADMIN_SECRET = os.getenv("TEST_ADMIN_SECRET", "123456").strip()
-
 MAX_DOCUMENT_SIZE_MB = 100
