+++ tests/e2e_test/conftest.py
@@ -175,11 +175,9 @@ def register_user():
     email = f"{username}@example.com"
     password = f"TestPwd!{random.randint(1000, 9999)}"
     data = {"username": username, "email": email, "password": password}
-    headers = {"X-Test-Secret": os.environ.get("TEST_ADMIN_SECRET", "").strip()}
     resp = httpx.post(
         f"{API_BASE_URL}/api/v1/test/register_admin",
         json=data,
-        headers=headers,
     )
     assert resp.status_code == HTTPStatus.OK, f"register failed: {resp.text}"
     user = resp.json()
